requirements.txt
hikari
hikari-lightbulb
python-dotenv
wikipedia
valorant
requests